move ports to shared
