+++ packages/client/src/commands.ts
@@ -9,12 +9,11 @@ import {
   window,
   workspace,
 } from 'vscode'
-import { DEFAULT_GAME_PORT, DEFAULT_REPL_PORT } from '../../server/src/settings'
 import {
   asShellString,
   fsToShell,
 } from './platform-string-utils'
-import { COMMAND_RUN_ALL_TESTS, COMMAND_RUN_GAME, COMMAND_RUN_PROGRAM, COMMAND_RUN_TEST, COMMAND_START_REPL, wollokLSPExtensionCode, COMMAND_INIT_PROJECT, COMMAND_DEBUG } from '../../shared/definitions'
+import { DEFAULT_GAME_PORT, DEFAULT_REPL_PORT, COMMAND_RUN_ALL_TESTS, COMMAND_RUN_GAME, COMMAND_RUN_PROGRAM, COMMAND_RUN_TEST, COMMAND_START_REPL, wollokLSPExtensionCode, COMMAND_INIT_PROJECT, COMMAND_DEBUG } from '../../shared/definitions'
 import { getLSPMessage } from './messages'
 
 export const subscribeWollokCommands = (context: ExtensionContext): void => {
